Fix parameter name typo, restore API surface
The real fix here is in lines 473/474, fixing a variable name collision issue. The extra parameterless method is to re-add the public API method you removed and which I was, in fact, using in a mod (which is how I found the naming issue in the first place - my mod simply stopped linking against the 1.20 DLLs).